meet/requirements.txt
# UNIO Backend Requirements

# Django Framework
Django==4.2.7
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Django Channels for WebSockets
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

# CORS Headers
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (optional)

# OAuth2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0

# Image Processing
Pillow==10.1.0

# Filtering
django-filter==23.5

# API Documentation (Swagger/OpenAPI)
drf-yasg==1.21.7

# Environment Variables
python-decouple==3.8

# ASGI Server
uvicorn==0.24.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development Tools
black==23.11.0
flake8==6.1.0
